Add support for attaching replies
Based on a config value, reply messages can now be attached to the work
item as files.
As part of this change, the strategy for attaching email messages to the
work item has changed a little bit, where the attachment file name will
indicate the date/time when it was attached (including for the original
message).
Also added code to delete the temporary attachment file once it has been
attached to the work item.