Add possibility to include sender of the e-mail in the comments
This introduces two new special values (##MessageLastReply and ##MessageLastReplyWithSender) and a new configuration option DefaultFieldValuesOnUpdate. DefaultFieldValuesOnUpdate is same as DefaultFieldValues except that its default values are used when an update is applied to a work item. ##MessageLastReplyWithSender allows to set the default value for field History similarly as is currently done for Description with ##MessageBody. This change also unifies the code a bit as we do not need to have special cases for History field. It should also allow to remove the configuration option OverrideChangedBy and replace it with ##Sender as default value for 'Changed By'.