Stricter match when using UseConversationGuidOnly (#86)
Mail2Bug does a StartsWith match to locate bugs based on the conversation ID, since the conversation IDs of two emails are not identical. However, when UseConversationGuidOnly is enabled, the IDs are of fixed length and we can do an equality match. This will also protect, to some degree, against false matches with inadvertent, manually entered values in the field.