Use a different URL for workitems. (#78)
The sample config uses a bug URL of the format _workitems#_a=edit&id=[_ID] and it works fine for the most parts. However, if the user is not already logged in, the bit after the # is lost after the login flow and the user ends up in the queries tab instead of the specific bug. The URL _workitems/edit/[_ID] seems to work fine in all cases.

The URL _workitems?_a=edit&id=[_ID] was also tried and works fine on VSTS, but is not recognized in TFS 2013.